+++ packages/compass-aggregations/src/components/focus-mode/focus-mode-stage-preview.tsx
@@ -12,8 +12,8 @@ import { PipelineOutputOptionsMenu } from '../pipeline-output-options-menu';
 import type { PipelineOutputOption } from '../pipeline-output-options-menu';
 import { connect } from 'react-redux';
 import type { RootState } from '../../modules';
-import { OutStagePreivew, MergeStagePreivew } from '../output-stage-preview';
-import { AtlasStagePreview } from '../atlas-stage-preview';
+import { OutStagePreivew, MergeStagePreivew } from '../stage-preview/output-stage-preview';
+import { AtlasStagePreview } from '../stage-preview/atlas-stage-preview';
 import { isMissingAtlasStageSupport } from '../../utils/stage';
 
 const containerStyles = css({
