+++ packages/compass-aggregations/src/components/stage-preview/atlas-stage-preview.tsx
