+++ packages/compass-aggregations/src/components/stage-preview/index.spec.tsx
@@ -4,13 +4,13 @@ import { render, screen } from '@testing-library/react';
 import { expect } from 'chai';
 import { Provider } from 'react-redux';
 
-import configureStore from '../stores/store';
+import configureStore from '../../stores/store';
 
-import { StagePreview } from './stage-preview';
+import { StagePreview } from './';
 import {
   MERGE_STAGE_PREVIEW_TEXT,
   OUT_STAGE_PREVIEW_TEXT,
-} from '../utils/stage';
+} from '../../utils/stage';
 
 const renderStagePreview = (
   props: Partial<ComponentProps<typeof StagePreview>> = {}
