+++ packages/compass-aggregations/src/components/stage-preview/index.tsx
@@ -12,10 +12,10 @@ import {
 } from '@mongodb-js/compass-components';
 import { Document } from '@mongodb-js/compass-crud';
 
-import type { RootState } from '../modules';
-import { isMissingAtlasStageSupport } from '../utils/stage';
+import type { RootState } from '../../modules';
+import { isMissingAtlasStageSupport } from '../../utils/stage';
 
-import LoadingOverlay from './loading-overlay';
+import LoadingOverlay from '../loading-overlay';
 import { AtlasStagePreview } from './atlas-stage-preview';
 import { OutStagePreivew, MergeStagePreivew } from './output-stage-preview';
