+++ packages/compass-aggregations/src/components/stage-preview/output-stage-preview.spec.tsx
